renaming file to normalize
