+++ src/js/components/chart/normalize.js
